+++ ft_putpointer.c
@@ -6,13 +6,13 @@
 /*   By: mboujama <mboujama@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/24 11:56:17 by mboujama          #+#    #+#             */
-/*   Updated: 2023/12/25 14:03:21 by mboujama         ###   ########.fr       */
+/*   Updated: 2023/12/26 10:32:55 by mboujama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_putpointer(unsigned int nb, int *len)
+void	ft_putpointer(unsigned long nb, int *len)
 {
 	ft_putstr("0x", len);
 	ft_puthexa(nb, 'x', len);
